requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.0
colorlog==6.8.0
comm==0.2.1
contourpy==1.2.0
cookiecutter==2.5.0
cycler==0.12.1
db-dtypes==1.2.0
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.101.0
flet==0.19.0
flet-core==0.19.0
flet-runtime==0.19.0
fonttools==4.47.0
git-lfs==1.6
google-api-core==2.15.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-bigquery==3.14.1
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
holidays==0.40
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
llvmlite==0.42.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numba==0.59.0
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
optuna==3.4.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pandas-gbq==0.20.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
prophet==1.1.5
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydata-google-auth==1.8.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
selenium==4.16.0
setuptools==69.0.3
shap==0.44.1
six==1.16.0
skforecast==0.11.0
slicer==0.0.7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
stanio==0.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.14.1
trio==0.23.2
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0.post1
websocket-client==1.7.0
websockets==11.0.3
flet_fastapi==0.19.0
